+++ sparkle/src/components/markdown/MentionBlock.tsx
