+++ sparkle/src/components/markdown/index.ts
@@ -2,7 +2,8 @@ export * from "./CiteBlock";
 export * from "./CodeBlock";
 export * from "./CodeBlockWithExtendedSupport";
 export * from "./ContentBlockWrapper";
+export * from "./ExtendedMarkdown";
+export * from "./LegacyMarkdown";
 export * from "./MarkdownCitation";
 export * from "./MarkdownContentContext";
 export * from "./MentionBlock";
-export * from "./RenderMessageMarkdown";
